feat(voice): one simulated voice sheet that rehearses the real Home feature (2.23.3)
Settings → Developer had two voice experiments: a "Voice input" sheet
that only transcribed speech, and a "Voice control" playground with its
own layout, a door-placement selector and a cancel-window stepper.
Neither resembled the Home voice control, which is the surface that
actually presses the door. Both are replaced by one "Simulated voice"
sheet that IS the Home feature, running against a pretend door.

Fidelity is structural, not a resemblance. The Home card body and its
recognizer plumbing move to a shared ui/voice/VoiceControlCard.kt
(VoiceControlCard + VoiceRecognizerEffects) rendered by BOTH surfaces,
so they cannot drift — there is only one of them. The surfaces differ in
exactly one place: which VoiceCommandEnvironment the ViewModel supplies.

So is the safety property. DefaultProfileViewModel has no
PushRemoteButtonUseCase at all, asserted by constructor reflection in
the new :androidApp SimulatedVoiceSafetyTest (mirrors the watch's
cannotReachTheRealRemoteButton; it lives in :androidApp because
:viewmodel's tests are commonTest, where JVM reflection is unavailable).

Removed with the playgrounds:
- VoiceExperimentState and its four ProfileViewModel members.
- The door-placement selector. The pretend door reacts to commands, so
  refusals stay reachable — open it and ask again for DOOR_ALREADY_OPEN,
  speak mid-transit for DOOR_MOVING. Only DOOR_STATE_UNKNOWN becomes
  unreachable here; the live Home surface reaches it via a stale
  check-in and VoiceCommandControllerTest covers it directly.
- The cancel-window stepper. The window is a settled fixed 3s
  everywhere, matching Home.
- Tap-to-copy of the structured verdict, which fed the eval corpus.
  docs/VOICE_COMMANDS.md records how to grow the corpus without it and
  how to add it back to the simulated sheet if that proves annoying.

Tests: three ProfileViewModelTest cases (loop end to end, already-open
refusal, cancel-on-leave) plus the safety test. Both new claims
mutation-verified — adding a real dependency to the forbidden set fails
the safety test, and no-op'ing voiceCommandBackgrounded fails
cancel-on-leave.

Patch: the live Home voice feature and the real-door path are untouched;
only developer-gated simulated surfaces changed.

validate.sh: 54 PASS / 0 FAIL.
