+++ MobileGarage/androidApp/src/main/java/com/chriscartland/garage/ui/ProfileContent.kt
@@ -17,15 +17,10 @@
 
 package com.chriscartland.garage.ui
 
-import android.app.Activity
-import android.content.ActivityNotFoundException
 import android.content.Intent
 import android.os.Build
-import android.speech.RecognizerIntent
 import android.widget.Toast
 import androidx.activity.compose.ReportDrawn
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Box
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
@@ -59,11 +54,10 @@ import com.chriscartland.garage.ui.settings.AccountBottomSheet
 import com.chriscartland.garage.ui.settings.AccountRowState
 import com.chriscartland.garage.ui.settings.NavRailBottomSheet
 import com.chriscartland.garage.ui.settings.SettingsContent
+import com.chriscartland.garage.ui.settings.SimulatedVoiceBottomSheet
 import com.chriscartland.garage.ui.settings.SnoozeBottomSheet
 import com.chriscartland.garage.ui.settings.SnoozeRowState
 import com.chriscartland.garage.ui.settings.VersionBottomSheet
-import com.chriscartland.garage.ui.settings.VoiceControlBottomSheet
-import com.chriscartland.garage.ui.settings.VoiceInputBottomSheet
 import com.chriscartland.garage.version.AppVersion
 import com.chriscartland.garage.viewmodel.ProfileViewModel
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
@@ -124,49 +118,15 @@ fun ProfileContent(
     var accountSheetOpen by rememberSaveable { mutableStateOf(false) }
     var versionSheetOpen by rememberSaveable { mutableStateOf(false) }
     var navRailSheetOpen by rememberSaveable { mutableStateOf(false) }
-    var voiceSheetOpen by rememberSaveable { mutableStateOf(false) }
-    var voiceControlSheetOpen by rememberSaveable { mutableStateOf(false) }
+    var simulatedVoiceSheetOpen by rememberSaveable { mutableStateOf(false) }
 
-    // Experimental voice-input playground (Settings → Developer → Voice
-    // input). The system speech dialog (RecognizerIntent) records and
-    // recognizes — this app never touches audio and needs no microphone
-    // permission. The transcript goes to the ViewModel's in-memory state
-    // and nowhere else; launching a new capture clears the previous one.
-    val voiceExperimentState by resolved.voiceExperimentState.collectAsState()
-    val voiceLauncher = rememberLauncherForActivityResult(
-        ActivityResultContracts.StartActivityForResult(),
-    ) { result ->
-        val transcript = if (result.resultCode == Activity.RESULT_OK) {
-            result.data
-                ?.getStringArrayListExtra(RecognizerIntent.EXTRA_RESULTS)
-                ?.firstOrNull()
-        } else {
-            null
-        }
-        resolved.reportVoiceExperimentTranscript(transcript)
-    }
-    val voicePrompt = stringResource(R.string.voice_experiment_prompt)
-    val onVoiceSpeakTap = {
-        resolved.clearVoiceExperiment()
-        val intent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
-            .putExtra(
-                RecognizerIntent.EXTRA_LANGUAGE_MODEL,
-                RecognizerIntent.LANGUAGE_MODEL_FREE_FORM,
-            ).putExtra(RecognizerIntent.EXTRA_PROMPT, voicePrompt)
-            .putExtra(RecognizerIntent.EXTRA_MAX_RESULTS, 1)
-        try {
-            voiceLauncher.launch(intent)
-        } catch (_: ActivityNotFoundException) {
-            resolved.reportVoiceExperimentUnavailable()
-        }
-    }
-
-    // Experimental voice-command playground (Settings → Developer →
-    // Voice control): the full command loop against a simulated door.
-    // Recognizer plumbing lives inside VoiceControlBottomSheet.
+    // Settings → Developer → Simulated voice: the Home tab's voice
+    // control rehearsed against a pretend door. Same controller, same
+    // gate, same card — only the environment differs, and this VM has
+    // no access to the real remote button. Recognizer plumbing lives
+    // inside SimulatedVoiceBottomSheet.
     val voiceCommandState by resolved.voiceCommandState.collectAsState()
     val voiceCommandDoorState by resolved.voiceCommandDoorState.collectAsState()
-    val voiceCommandArmedWindowMs by resolved.voiceCommandArmedWindowMs.collectAsState()
 
     // TTL-gated revalidate, once per screen entry (STATUS_CACHE_PLAN.md
     // D3). The cached snooze state renders instantly (hydrated across
@@ -281,8 +241,7 @@ fun ProfileContent(
             onDiagnosticsTap = onNavigateToDiagnostics,
             onLayoutDebugChange = resolved::setLayoutDebugEnabled,
             onNavRailTap = { navRailSheetOpen = true },
-            onVoiceInputTap = { voiceSheetOpen = true },
-            onVoiceControlTap = { voiceControlSheetOpen = true },
+            onSimulatedVoiceTap = { simulatedVoiceSheetOpen = true },
         )
         SnackbarHost(
             hostState = snackbarHostState,
@@ -335,53 +294,15 @@ fun ProfileContent(
         )
     }
 
-    if (voiceSheetOpen) {
-        VoiceInputBottomSheet(
-            state = voiceExperimentState,
-            onSpeakTap = onVoiceSpeakTap,
-            onCopy = { label, value ->
-                clipboardManager.setText(AnnotatedString(value))
-                // Android 13+ shows the OS clipboard chip after setText;
-                // an app Toast on top is duplicate noise (same gate as
-                // the Version sheet).
-                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
-                    Toast
-                        .makeText(
-                            context,
-                            resources.getString(R.string.profile_version_toast_copied, label),
-                            Toast.LENGTH_SHORT,
-                        ).show()
-                }
-            },
-            onDismiss = { voiceSheetOpen = false },
-        )
-    }
-
-    if (voiceControlSheetOpen) {
-        VoiceControlBottomSheet(
+    if (simulatedVoiceSheetOpen) {
+        SimulatedVoiceBottomSheet(
             state = voiceCommandState,
             doorState = voiceCommandDoorState,
-            armedWindowMs = voiceCommandArmedWindowMs,
             onMicTap = resolved::voiceCommandMicTap,
             onTranscript = resolved::voiceCommandTranscript,
             onCaptureUnavailable = resolved::voiceCommandCaptureUnavailable,
             onBackgrounded = resolved::voiceCommandBackgrounded,
-            onSetDoorState = resolved::setVoiceCommandDoorState,
-            onSetArmedWindowMs = resolved::setVoiceCommandArmedWindowMs,
-            onCopy = { label, value ->
-                clipboardManager.setText(AnnotatedString(value))
-                // Same Android 13+ gate as the Version sheet: the OS
-                // clipboard chip already confirms the copy there.
-                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
-                    Toast
-                        .makeText(
-                            context,
-                            resources.getString(R.string.profile_version_toast_copied, label),
-                            Toast.LENGTH_SHORT,
-                        ).show()
-                }
-            },
-            onDismiss = { voiceControlSheetOpen = false },
+            onDismiss = { simulatedVoiceSheetOpen = false },
         )
     }
