--- MobileGarage/androidApp/src/main/java/com/chriscartland/garage/ui/settings/VoiceInputBottomSheet.kt
@@ -1,273 +1,0 @@
-/*
- * Copyright 2026 Chris Cartland. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.chriscartland.garage.ui.settings
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Mic
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FilledTonalButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.chriscartland.garage.R
-import com.chriscartland.garage.domain.model.VoiceIntent
-import com.chriscartland.garage.domain.model.VoiceIntentClassification
-import com.chriscartland.garage.domain.model.VoiceIntentConfidence
-import com.chriscartland.garage.viewmodel.VoiceExperimentState
-
-/**
- * Experimental voice-input playground (Settings → Developer → Voice
- * input). One tap on Speak launches the system speech prompt
- * (`RecognizerIntent` — the launch lives in `ProfileContent`, which
- * owns the activity-result plumbing); the transcript renders here and
- * nowhere else. Deliberately NOT wired to any door action, and the
- * state is ViewModel-memory only — leaving the screen discards it.
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun VoiceInputBottomSheet(
-    state: VoiceExperimentState,
-    onSpeakTap: () -> Unit,
-    onCopy: (label: String, value: String) -> Unit,
-    onDismiss: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = sheetState,
-        modifier = modifier,
-    ) {
-        VoiceInputSheetContent(
-            state = state,
-            onSpeakTap = onSpeakTap,
-            onCopy = onCopy,
-        )
-    }
-}
-
-/**
- * Sheet content extracted so previews can render it directly (the
- * ModalBottomSheet show animation doesn't run under previews).
- */
-@Composable
-fun VoiceInputSheetContent(
-    state: VoiceExperimentState,
-    onSpeakTap: () -> Unit,
-    onCopy: (label: String, value: String) -> Unit = { _, _ -> },
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .verticalScroll(rememberScrollState())
-            .padding(horizontal = 24.dp, vertical = 16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        Text(
-            text = stringResource(R.string.voice_experiment_sheet_title),
-            style = MaterialTheme.typography.titleLarge,
-        )
-        Text(
-            text = stringResource(R.string.voice_experiment_disclaimer),
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        FilledTonalButton(
-            onClick = onSpeakTap,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Mic,
-                contentDescription = null,
-                modifier = Modifier.size(20.dp),
-            )
-            Text(
-                text = stringResource(R.string.voice_experiment_speak_button),
-                modifier = Modifier.padding(start = 8.dp),
-            )
-        }
-        when (state) {
-            VoiceExperimentState.Idle -> Text(
-                text = stringResource(R.string.voice_experiment_hint),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            is VoiceExperimentState.Transcript -> Column(
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                val transcriptLabel = stringResource(R.string.voice_copy_label_transcript)
-                val verdictLabel = stringResource(R.string.voice_copy_label_verdict)
-                // Tap the transcript to copy the raw text.
-                Surface(
-                    color = MaterialTheme.colorScheme.surfaceContainerHighest,
-                    shape = MaterialTheme.shapes.medium,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clip(MaterialTheme.shapes.medium)
-                        .clickable { onCopy(transcriptLabel, state.text) },
-                ) {
-                    Text(
-                        text = state.text,
-                        style = MaterialTheme.typography.bodyLarge,
-                        modifier = Modifier.padding(16.dp),
-                    )
-                }
-                // Tap the verdict to copy the structured summary
-                // (input + intent + confidence + engine) for pasting
-                // into a discussion.
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(4.dp),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clip(MaterialTheme.shapes.medium)
-                        .clickable { onCopy(verdictLabel, state.clipboardSummary()) },
-                ) {
-                    Text(
-                        text = stringResource(
-                            R.string.voice_experiment_classification,
-                            state.classification.intent.displayName(),
-                            state.classification.confidence.displayName(),
-                        ),
-                        style = MaterialTheme.typography.titleMedium,
-                    )
-                    Text(
-                        text = stringResource(
-                            R.string.voice_experiment_engine,
-                            state.engineName,
-                        ),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-                Text(
-                    text = stringResource(R.string.voice_experiment_copy_hint),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-            VoiceExperimentState.NoSpeech -> Text(
-                text = stringResource(R.string.voice_experiment_no_speech),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            VoiceExperimentState.Unavailable -> Text(
-                text = stringResource(R.string.voice_experiment_unavailable),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.error,
-            )
-        }
-    }
-}
-
-// User-visible names for the classification result. UI-local mapping —
-// the domain enums stay display-free.
-@Composable
-private fun VoiceIntent.displayName(): String =
-    when (this) {
-        VoiceIntent.OPEN -> stringResource(R.string.voice_intent_open)
-        VoiceIntent.CLOSE -> stringResource(R.string.voice_intent_close)
-        VoiceIntent.UNKNOWN -> stringResource(R.string.voice_intent_unknown)
-    }
-
-@Composable
-private fun VoiceIntentConfidence.displayName(): String =
-    when (this) {
-        VoiceIntentConfidence.HIGH -> stringResource(R.string.voice_confidence_high)
-        VoiceIntentConfidence.MEDIUM -> stringResource(R.string.voice_confidence_medium)
-        VoiceIntentConfidence.NONE -> stringResource(R.string.voice_confidence_none)
-    }
-
-// `private` so `checkPreviewCoverage` exempts them (same rationale as
-// NavRailBottomSheet: developer-gated sheet verified on a real device;
-// these are Android Studio references only).
-@Preview
-@Composable
-private fun VoiceInputSheetContentIdlePreview() {
-    Surface {
-        VoiceInputSheetContent(
-            state = VoiceExperimentState.Idle,
-            onSpeakTap = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun VoiceInputSheetContentTranscriptPreview() {
-    Surface {
-        VoiceInputSheetContent(
-            state = VoiceExperimentState.Transcript(
-                text = "open the garage door",
-                classification = VoiceIntentClassification(
-                    intent = VoiceIntent.OPEN,
-                    confidence = VoiceIntentConfidence.HIGH,
-                ),
-                engineName = "Rules v3",
-            ),
-            onSpeakTap = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun VoiceInputSheetContentUnknownIntentPreview() {
-    Surface {
-        VoiceInputSheetContent(
-            state = VoiceExperimentState.Transcript(
-                text = "what a nice day",
-                classification = VoiceIntentClassification(
-                    intent = VoiceIntent.UNKNOWN,
-                    confidence = VoiceIntentConfidence.NONE,
-                ),
-                engineName = "Rules v3",
-            ),
-            onSpeakTap = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun VoiceInputSheetContentNoSpeechPreview() {
-    Surface {
-        VoiceInputSheetContent(
-            state = VoiceExperimentState.NoSpeech,
-            onSpeakTap = {},
-        )
-    }
-}
