+++ MobileGarage/viewmodel/src/commonMain/kotlin/com/chriscartland/garage/viewmodel/ProfileViewModel.kt
@@ -32,7 +32,6 @@ import com.chriscartland.garage.domain.model.NavigationRailLayout
 import com.chriscartland.garage.domain.model.SnoozeAction
 import com.chriscartland.garage.domain.model.SnoozeDurationUIOption
 import com.chriscartland.garage.domain.model.SnoozeState
-import com.chriscartland.garage.domain.model.VoiceIntentClassification
 import com.chriscartland.garage.domain.model.WatchAppStatus
 import com.chriscartland.garage.domain.model.WatchInstallAction
 import com.chriscartland.garage.domain.model.WatchInstallResult
@@ -62,49 +61,6 @@ import kotlinx.coroutines.launch
 
 private const val SNOOZE_ACTION_RESET_DELAY_MS = 10_000L
 private const val WATCH_INSTALL_ACTION_RESET_DELAY_MS = 10_000L
-
-/**
- * State of the experimental voice-input playground (Settings →
- * Developer → Voice input). Deliberately in-memory only: it lives in
- * the screen ViewModel, so it survives rotation but evaporates when the
- * user leaves the screen or the process dies. Nothing is written to
- * DataStore/Room and no door action is ever taken from it.
- */
-sealed interface VoiceExperimentState {
-    /** Nothing captured yet, or cleared by starting a new capture. */
-    data object Idle : VoiceExperimentState
-
-    /**
-     * The recognizer returned a transcript, classified into a door
-     * intent + confidence by the DI-selected engine. Classification is
-     * display-only in this experiment — nothing acts on it.
-     */
-    data class Transcript(
-        val text: String,
-        val classification: VoiceIntentClassification,
-        val engineName: String,
-    ) : VoiceExperimentState {
-        /**
-         * Paste-friendly structured summary for sharing/discussing
-         * classifier results. Uses raw enum names (locale-independent,
-         * stable across app versions) and quotes the input so leading/
-         * trailing whitespace is visible.
-         */
-        fun clipboardSummary(): String =
-            """
-            input: "$text"
-            intent: ${classification.intent.name}
-            confidence: ${classification.confidence.name}
-            engine: $engineName
-            """.trimIndent()
-    }
-
-    /** The capture ended without usable speech (silence, cancel, no match). */
-    data object NoSpeech : VoiceExperimentState
-
-    /** No speech recognizer exists on this device. */
-    data object Unavailable : VoiceExperimentState
-}
 
 /**
  * Drives the Settings screen — one VM per screen (ADR-026). Aggregates the
@@ -179,41 +135,24 @@ interface ProfileViewModel {
     val watchInstallAction: StateFlow<WatchInstallAction>
 
     /**
-     * Experimental voice-input playground state. See
-     * [VoiceExperimentState] — in-memory only, never persisted, never
-     * wired to any action.
-     */
-    val voiceExperimentState: StateFlow<VoiceExperimentState>
-
-    /**
-     * Call when a new capture starts: clears the previous transcript so
-     * each prompt replaces the last one.
-     */
-    fun clearVoiceExperiment()
-
-    /**
-     * Report the recognizer outcome. Null or blank text means no usable
-     * speech was captured.
-     */
-    fun reportVoiceExperimentTranscript(text: String?)
-
-    /** Report that this device has no speech recognizer at all. */
-    fun reportVoiceExperimentUnavailable()
-
-    /**
-     * Experimental voice-command UX playground (Settings → Developer →
-     * Voice control): the full mic → classify → gate → cancel-window →
-     * commit loop, running against a simulated door and a pretend
-     * button. The real door is never touched. State machine:
+     * Settings → Developer → Simulated voice: the Home tab's voice
+     * control, rehearsed safely. The same mic → classify → gate →
+     * cancel-window → commit loop and the same 3s cancel window as
+     * Home, but the commit presses a pretend button on a pretend door
+     * ([SimulatedVoiceCommandEnvironment]). The real door cannot be
+     * reached from here — this ViewModel has no
+     * `PushRemoteButtonUseCase`. State machine:
      * [VoiceCommandController] in `:usecase`.
      */
     val voiceCommandState: StateFlow<VoiceCommandState>
 
-    /** The simulated door the playground's gate checks against. */
+    /**
+     * The pretend door the gate checks against. Displayed read-only:
+     * refusals ("already open", "the door is moving") are unreadable
+     * without it, and it reacts to committed commands the way the real
+     * one does.
+     */
     val voiceCommandDoorState: StateFlow<VoiceDoorState>
-
-    /** Cancel-window length, user-adjustable within the clamp range. */
-    val voiceCommandArmedWindowMs: StateFlow<Long>
 
     /** Mic tap: always starts over; cancels a pending command first. */
     fun voiceCommandMicTap()
@@ -225,15 +164,10 @@ interface ProfileViewModel {
     fun voiceCommandCaptureUnavailable()
 
     /**
-     * The playground left the screen: cancels a pending (Armed) command
+     * The surface left the screen: cancels a pending (Armed) command
      * so nothing commits off-screen.
      */
     fun voiceCommandBackgrounded()
-
-    /** Place the simulated door directly to exercise gate paths. */
-    fun setVoiceCommandDoorState(state: VoiceDoorState)
-
-    fun setVoiceCommandArmedWindowMs(ms: Long)
 
     fun signInWithGoogle(idToken: GoogleIdToken)
 
@@ -321,13 +255,11 @@ class DefaultProfileViewModel(
     private val _watchInstallAction = MutableStateFlow<WatchInstallAction>(WatchInstallAction.Idle)
     override val watchInstallAction: StateFlow<WatchInstallAction> = _watchInstallAction
 
-    private val _voiceExperimentState =
-        MutableStateFlow<VoiceExperimentState>(VoiceExperimentState.Idle)
-    override val voiceExperimentState: StateFlow<VoiceExperimentState> = _voiceExperimentState
-
-    // Voice-command playground: the real state machine wired to a fake
-    // world (simulated door + pretend button). viewModelScope-owned so
-    // the experiment's state evaporates with the screen VM.
+    // Simulated voice: the real state machine wired to a fake world
+    // (pretend door + pretend button). viewModelScope-owned so the
+    // rehearsal's state evaporates with the screen VM. The cancel window
+    // is left at the default, which is the same 3s Home commits on — the
+    // point of this surface is to behave exactly like the live one.
     private val voiceCommandEnvironment = SimulatedVoiceCommandEnvironment(scope = viewModelScope)
     private val voiceCommandController = VoiceCommandController(
         classify = classifyVoiceIntentUseCase,
@@ -336,7 +268,6 @@ class DefaultProfileViewModel(
     )
     override val voiceCommandState: StateFlow<VoiceCommandState> = voiceCommandController.state
     override val voiceCommandDoorState: StateFlow<VoiceDoorState> = voiceCommandEnvironment.doorState
-    override val voiceCommandArmedWindowMs: StateFlow<Long> = voiceCommandController.armedWindowMs
 
     // Cached so the snooze action can attach the latest door change time
     // without the UI having to thread it through.
@@ -382,26 +313,6 @@ class DefaultProfileViewModel(
         }
     }
 
-    override fun clearVoiceExperiment() {
-        _voiceExperimentState.value = VoiceExperimentState.Idle
-    }
-
-    override fun reportVoiceExperimentTranscript(text: String?) {
-        _voiceExperimentState.value = if (text.isNullOrBlank()) {
-            VoiceExperimentState.NoSpeech
-        } else {
-            VoiceExperimentState.Transcript(
-                text = text,
-                classification = classifyVoiceIntentUseCase(text),
-                engineName = classifyVoiceIntentUseCase.engineName,
-            )
-        }
-    }
-
-    override fun reportVoiceExperimentUnavailable() {
-        _voiceExperimentState.value = VoiceExperimentState.Unavailable
-    }
-
     override fun voiceCommandMicTap() = voiceCommandController.onMicTap()
 
     override fun voiceCommandTranscript(text: String?) = voiceCommandController.onTranscript(text)
@@ -409,10 +320,6 @@ class DefaultProfileViewModel(
     override fun voiceCommandCaptureUnavailable() = voiceCommandController.onCaptureUnavailable()
 
     override fun voiceCommandBackgrounded() = voiceCommandController.onBackgrounded()
-
-    override fun setVoiceCommandDoorState(state: VoiceDoorState) = voiceCommandEnvironment.setDoorState(state)
-
-    override fun setVoiceCommandArmedWindowMs(ms: Long) = voiceCommandController.setArmedWindowMs(ms)
 
     override fun signInWithGoogle(idToken: GoogleIdToken) {
         viewModelScope.launch(dispatchers.io) {
